BUG#35257 - Falcon: tablespace names are case sensitive
Fixed that falcon tablespace names were case sensitive.
With this fix the rules for tablespace names are the same
as for table names, that is case insensitive.

Fixed by performing case up on tablespace name whenever
tablespace is created/dropped or a table is created/altered.

Note that it is incompatible fix:
- it is not possible to drop existing tablespace created
  by previous versions, if it's name wasn't in upper case.
- it is not possible to create tables using tablespace
  created by previous versions, if tablespace name wasn't
  in upper case.