Bug #37714 rpl.rpl_heartbeat fails sporadically in pushbuild due to timeout
The reason of the failure on windows platform was not detected. Still, a piece of heartbeat code
had a flaw fixed with Bug #39077.
It's probable that the patch for the latter bug, which is going to be pushed to 6.0 main,
can help with the current.

Attempted to fix with patch for Bug #39077.