BUG#40434 : Replication should not be allowed to start if restore is running
Currently, the work of WL#4209 and WL#4280 provide mechanisms to control 
how replication and backup interact. However, we also need to have a method 
to prevent replication from starting on a slave if a restore is ongoing on that
same slave.

This patch prohibits a slave from starting replication when a restore is in
progress.