Bug #41545 sporadic pushbuild failure: [Falcon] Error: read error on database header
Falcon crashes in ASSERT if it cannot read header from tablespace. The actual problem
seems to be the lack of diskspace on the pushbuild box.

Solution: throw exception instead of ASSERT. Exceptions are caught within plugin initialization code, i.e errors like this are handled gracefully 
with error message instead of just aborting the server.