Bug #40305 - Backup: Bad feedback when BACKUP/RESTORE operation aborted.
This patch refactors the code reporting termination of BACKUP/RESTORE statement. 
After the changes "Backup/Restore completed" is shown only if operation has been 
completed. In case of abnormal termination (error, interruption) a warning 
"Operation aborted" or "Operation aborted - data might be corrupted" is logged. 
The latter is in the case of RESTORE operation which has modified some data in 
the server.

The main changes are:
- Function backup::Logger::report_stop() is replaced by 2 new functions: 
  report_completed() and report_aborted() to differentiate the feedback given in 
  the two situations.
- Make sure that a call to report_completed() or report_aborted() always matches 
  a previous call to report_start(). This is done in Backup_restore_ctx::close() 
  which is called from the destructor.
- Change the logic for removing backup file (in case BACKUP operation is not 
  completed). Before this was guarded by explicit m_remove_loc flag. Now we can 
  directly check if BACKUP operation has been completed successfully (i.e. 
  report_completed() was called) or not.

Note: the new functionality is covered by existing test cases.