Bug#40814 CSV engine does not parse \X characters when they occur in unquoted fields
When a .CSV file for table in the CSV engine contains
\X characters as part of unquoted fields, e.g.
      
2,naraya\nan
      
\n is not interpreted as a new line (it is however interpreted as a
newline in a quoted field).
      
The old algorithm copied the entire value for a unquoted field without
parsing the \X characters. 
      
The new algorithm adds the capability to handle \X characters in the 
unquoted fields of a .CSV file.