BUG#40116: Uncommited changes are replicated and stay on slave after rollback on master
Post-merge fixes to disable part of a test that is not
relevant any more.

When replicating from MyISAM to any transactional table
with AUTOCOMMIT=0 and not using a COMMIT to terminate
the transaction, the changes will not be applied to the
slave since they are not committed to the transactional
engine.