This patch contains fixes for two bugs:
bug#40607 Unsigned indexes broken
      
bug#41582 Index access returns no result for indexes with numeric
data type
------------------------------------------------------------------
      
These bugs were so close to each other that committing the fixes in a
single patch seems most reasonable. Both bugs happened because Falcon,
when extracting the key value of a key (in
StorageDatabase::getSegmentValue), was unable to separate keys for the
decimal type from keys for other types also mapping to the
HA_KEYTYPE_BINARY key type.
      
In the case of #40607 this would result in index keys for the unsigned
numeric/decimal type to be treated the same way as the keys for
(f. ex) years (represented as unsigned tiny ints) although their key
format differs. In the case of #41582 this would lead to a numeric/decimal key with a
precision less than 19 to be interpreted the same was as an 8 byte
bigint.

I have extended my previous patch to this bug. The approach taken is
to augment the struct "StorageSegment" in StorageTableShare.h with an
enum that specifies the format of the key. The format is an 'ad hoc'
format and is set in StorageInterface::getKeyDesc, based on the key
type and the field type. For most cases, there is now a 1:1 relation
between the switch-cases in StorageInterface::getKeyDesc and the
switch-cases in StorageDatabase::getSegmentValue. The exception is for
HA_KEYTYPE_BINARY, where we now separate between binary integers,
binary strings and binary newdecimal. We have also made an exception
for HA_KEYTYPE_ULONG_INT, where values for timestamps should be
multiplied by 1000.  This way, most of the logic is moved from create
time to run time, and we get a correct separation of the different key
formats.

I have also added test cases to test the bugs. These are unmodified
since my previous commit, and have already been reviewed by John.
      
      
Modified file 'storage/falcon/StorageTableShare.h'
--------------------------------------------------
Added the KeyFormat enum definition. Augmented the StorageSegment
struct to include the enum (keyFormat)
      
      
Modified file 'storage/falcon/ha_falcon.cpp'
-------------------------------------------- 
Set the key format depending on the field type and key type. This is
done in StorageInterface::getKeyDesc. This is based upon a study of
the Field class hierarchy in sql/field.cc and sql/field.h
      
      
Modified file 'storage/falcon/StorageDatabase.cpp'
-------------------------------------------------- 
Case on the key format and extract the key values. The only real
difference here is for fields with a key type of HA_KEYTYPE_BINARY and
HA_KEYTYPE_ULONG_INT
      
      
Added file 'mysql-test/suite/falcon/t/falcon_bug_40607.test'
------------------------------------------------------------
Added a test testing that the patch fixes #40607. The test is based on
the bug report
      
      
Added file 'mysql-test/suite/falcon/r/falcon_bug_40607.result'
--------------------------------------------------------------
The expected test result
      
      
Added file 'mysql-test/suite/falcon/t/falcon_bug_41582.test'
------------------------------------------------------------
Added a test testing that the patch fixes #41582. Tests for the decimal
type, then for the numeric type, which both map the same internal
type. Uses values with a precision greater than 9, because these will
not fit within a group of 4 bytes. Also checks some numbers with a
scale different than 0 just to be on the safe side
      
Added file 'mysql-test/suite/falcon/r/falcon_bug_41582.result'
--------------------------------------------------------------
The expected test result