Fix for Bug#40128: drop-no_root fails under windows in 6.0-rpl.
The problem was that directories with no permission (000) files
are deleted differently on UNIX and on Windows.

On UNIX, 000-permission file is deleted perfectly, but other files
are left in the directory.

On Windows, 000-permission file is not deleted, but other files
are deleted.

Also, the fix needed a change in mysqltest.c: 'chmod' directive
should return a positive error code (in order to be handled).
It's decided to return a constant '1' for all error codes
just to be OS-independent.