Fixing sporadic failures of rpl_backup test in PB.
When running rpl_backup on slow platforms (e.g. bp-valgrind) 
sometimes the output of SHOW SLAVE STATUS was not as expected. 
This is most probably due to the fact that even if 
wait_for_slave_to_start.inc script detects that both IO 
and SQL threads are running, things which can change status 
of a slave still can happen. 

In this case it was the value of Relay_Master_Log_File parameter 
which sometimes was empty instead of being initialized.

This patch fixes the problem by masking Relay_Master_Log_File 
which is not really relevant.