This is a fix for Bug#42196 Issue with key look up for bit field in Falcon table after insert + stored func
This bug was already partly fixed by the patch for #40607 and #41582,
but some additional cleaning and testing is necessary. The problem was
that the old implementation of StorageDatabase::getSegmentValue would
choose code paths depending on 'segment->isUnsigned' when extracting
the value from keys of type HA_KEYTYPE_BINARY. 'segment->isUnsigned'
was set in StorageInterface::getKeyDesc in the following way, when
creating the index:

    segment->isUnsigned = (part->field->flags & ENUM_FLAG) ?
            true : ((Field_num*) part->field)->unsigned_flag;

The BIT type (class Field_bit_as_char) is not a subclass of Field_num,
which this is casted to, and has no 'unsigned_flag'. For BITs,
segment->isUnsigned would therefore contain more or less random
data. In the cases where this random data was '0', the bug would
occur.

The previously mentioned patch removed the need to check on
segment->isUnsiged. This patch removes the flag from the
StorageSegment altogether, as well as adding a test for the bug.



Modified file 'storage/falcon/StorageTableShare.h'
--------------------------------------------------
Removed 'isUnsigned' from the StorageSegment


Modified file 'storage/falcon/ha_falcon.cpp'
--------------------------------------------------
Removed the usage of 'segment->isUnsigned'


Added file 'mysql-test/suite/falcon/t/falcon_bug_42196.test'
--------------------------------------------------------------
Added a test for the bug, based on the report


Added file 'mysql-test/suite/falcon/r/falcon_bug_42196.result'
--------------------------------------------------------------
The expected output of the test