Bug#42136: Falcon: odd behavior for range query on index prefix when MRR is disabled
Problem:
Falcon does not search efficiently for prefix queries (multi-segment indexes). Index scanning  produces a lot of false positives, that are filtered out in the post-processing step, which negatively affects performance.

Analysis:
In the bug description, keys are numeric and and Falcon is asked to retrieve all keys where the first segment corresponds to 2. Falcon would construct a prefix 0x02c0 (0x02 is segment byte. 0xc0 is encoded 2) and use this prefix for scanning. 
But that  prefix is too short - keys build from integers in the [2, 131071] range will all start with the same prefix 0x02c0.

Solution:
Since we are using multi-segment keys for search, we can construct a better prefix. Taking into account rules for encoding of multi-segment key, we can padd our key with 0x0 to the length that is multiple of 6, and add segment  byte for the next segment. So instead of searching for "0x02c0" prefix we will be be searching for "0x02c00000000001" which yields a much better result.