Bug #42414 falcon_bug_33211 fails when trying to create a table in a non-existing table space.
This is a regression introduced with the fix for #42392  !Falcon bugs in recovery of create tablespace". The root cause is that tablespace bootstrap does not try open all tablespaces anymore. The code in Falcon was not prepared to handle inactive tablespace in this specific case 
(create table in a tablespace)

Fix:  open tablespace if it is inactive, regardless if it is an open open by id or by name.