Bug #42366: server-cert.pem expired: "Not After : Jan 27 08:54:13 2009 GMT"
Re-generated the PKI files needed.
Removed the ones that are not needed.
Updated the tests to reference the correct SSL subject.

Note: This patch is written by Joro. It is applied to 6.0-backup-merge branch to quickly fix PB errors in mysql-6.0 since 6.0-bugteam branch is not ready for merge into main.