Removing non-determinism from backup_vp_* tests.
In these tests data is inserted into table by threads executing
in parallel and thus the order of rows in the table can be 
non-deterministic. This made test fail sometimes.

This patch fixes the order in which table contents is shown in 
result file by adding ORDER BY clauses to SELECT statements.