BUG#42444: main.change_user: result from com_select differs between embedded and other runs
In 6.0-rpl this test was failing when mtr was running tests in embedded mode.
      
This happened because when in embedded mode, mtr does run "check_testcase" which
seems to cause extra selects. Since this test case relies on:

SHOW GLOBAL STATUS LIKE 'com_select';

for passing, this will get different values whether mtr is run in embedded 
mode or other modes, therefore the test failure surfaces.
      
This patch address this issue by making the part of the test dependent from 
SHOW GLOBAL STATUS LIKE 'com_select'; to work with relative instead of 
absolute values (have a test start timestamp and then subtract it in 
the checks).