Bug #42448 assertion in IO::writePage in recovery
Problem:
If a log record SRLSectionPage is processed , redoSectionPage assumes it needs to create a new Section page. It is not always true, pageNumber 0 indicates section will be deleted. Currently, redoSectionPage will create a new section page 0 and mark it dirty. However, page number 0 is  reserved for tablespace header, and IO::writePage has corresponding assertion for it. recovery will crash in this assertion ,once newly created  section page 0 is written to disk.

Fix : do not create section page 0 in recovery.