Bug#41227  Falcon: potential corruption of RecordLocatorPage on recovery
Problem: If falcon pager size is set to 32K, recovery 
involving blob updates will crash in assertion. 

Section::redoBlobUpdate and Section::redoBlobDelete will
try to set available space to the page size (32K)in record locator page
However, 32K=32768  becomes negative (-32768) when converted to short.
RecordLocatorPage::linkSpaceSlot() checks for positive value for 
"space available" and asserts.

Fix is to set spaceAvailable to more realistic value 
(page size - page header overhead) that is guaranteed not to overflow 
and not to change sign of a short even for 32 K pages.