Bug #42542 falcon_bug_28095 crashes in StorageHandler::dropTempTables
The reason for the crash is accessing Dbb::sections , which is NULL pointer in this case (temp tablespace is not opened during recovery)

Solution:
Make Dbb::sections an array instead of pointer, that  is initialized only when tablespace is created or opened. Initialize this array with NULLs in Dbb constructor.

No test case provided,  it seems quite tricky to create one, we should get a serial log that does not try to open temp tablespace during recovery.
(checkpoint or switching log files twice would do the trick).