Bug#42227 : Test falcon.falcon_bug_28095 fails sporadically on pushbuild
The reason for test failure is a memory leak found by valgrind.
Falcon has THR_LOCK in its StorageTalbeShareStructure ( dubbed "impure"). The lock is initialized when table is opened, but it is  normally not destroyed (only when table is deleted). In the failing test case valgrind notices the leak, because table is created  and not deleted before shutdown.

Fix is to destroy the THR_LOCK  when StorageTableShare is destroyed.