Bug #40975 - Database restored through named pipes brings up a corrupt database.
The problem was buried deep inside the backup stream library code. Inside 
bstream_read_part() function (in backup/stream_v1_transpor.c) there is a branch 
which reads bytes directly from an underlying stream to the provided input 
buffer (the old code):

  else
  {
    /* read directly into data area */
    ASSERT(s->buf.header > s->buf.pos);

    howmuch= data->end - data->begin;
    if ((s->buf.pos + howmuch) > s->buf.header)
      howmuch= s->buf.header - s->buf.pos;

    saved= *data;
    data->end= data->begin + howmuch;

    ret= as_read(&s->stream, data, buf);
    if (ret == BSTREAM_ERROR)
      return SERROR(s);
    if (ret == BSTREAM_EOS)
      s->state= EOS;

    s->buf.begin += data->begin - saved.begin;
    s->buf.pos= s->buf.begin;

    data->begin= data->end;
    data->end= saved.end;
  }

After determining how much data can be read (howmuch variable) the data blob 
describes the memory area where the data should be stored. as_read() function 
reads bytes from the underlying stream to the area described by data and updates 
data blob to describe the space which was *not* filled with bytes:

     |- the original data blob |
     |                         |
  ---[***************[---------]------
     |               |         |
     |- bytes read --|- data --|
        from stream     blob after the operation

In the case depicted above, when not all the available area was filled with 
bytes from the stream and data blob is non-empty, the assignment 

 data->begin= data->end

incorrectly modifies data->begin pointer which in this situation is different 
from data->end. This causes further confusion in the internal buffering code.

The assignment is completely wrong and serves no purpose here - the patch 
removes it. This problem was not detected before due to the fact that in the 
usual situation the whole available area is filled with data from the stream so 
that after as_read() data->begin == data->end. In this situation the assignment 
has no effect. However, when reading from a pipe, not always enough bytes is 
available in the stream to fill the buffer and above situation can happen.

In the test we trigger the errornous situation using error injection code.