New test related to Bug#41870 and Bug#42202, verifying that scheduled scavenging of old record versions in Falcon works with BLOBs, i.e. that the tablespace does not grow indefinitely when updating a BLOB multiple times.
This test is perhaps more of a functional test of timing (schedule) based scavenging 
related to BLOB updates than a real regression test for bug 41870 (Unbounded 
tablespace growth when updating BLOB record). This is because in order to be as 
timing-independent as possible, the test requires a fixed non-default value of 
the option falcon_scavenge_schedule (currently set to 'every 5 seconds').

On a standard disk based system the test may take a long time (~2 min on a relatively
fast server with backround load) to run, and requires 30-50 MB of space, so this is
a --big-test. In memory (--mem) the test should normally run in about 6-7 seconds.