Bug #42560 Falcon recovery error:  wrong page type, expected 5 got <something>
Problem:
during index creation root page number is not logged. 
The page is not recreated on recovery which may result 
in corrupted index.


Fix:
- Log index root page number and recreate index root
page in recovery of SRLCreateIndex. Do it in phase 2
of recovery, instead of "redo" phase.
(phase2 is responsible for recreating "physical" objects
, e.g pages)

- Also, SRLDeleteIndex will be executed in phase 2 of recovery

- Log index root page number when creating index and 
recreate this page on recovery. Refactor code so 
createIndexRoot can be called inside and outside recovery.
The code is more compact and better readable.

-  Simplify redoIndexPage.
Do not care about left and right neighbour pages.
they are logged separately during splits, and there is no 
need for careful trialReads.

- Cleanup: 
 remove misleadingly named  IndexPage::createNewLevel(), 
   substitute with initRootPage().

 remove setIndexRoot(), its single usage was within 
 redoDeleteIndex()