Since the last commit, I have added some comments to the code and hopefully fixed the poor indentation in this commit message
#34478 and #33190:
------------------

I will base the explanation on the following
insert statement:

insert into t values (0x00),(0x0000),(0x0000),(0x00);

DeferredIndex::addNode in DeferredIndex.cpp (way down the call stack
from StorageInterface::write_row) adds one index node at a time for
each inserted value. It compares the index keys and determines in
which order the nodes should be arranged, in an array of nodes
(leaf->nodes). After inserting the three first values, 0x00, 0x0000
and 0x0000, the array looks like this

[0x00|0x0000|0x0000|...] 

which is correct. 

When inserting the forth value, DeferredIndex::compare(DINode* node1, DINode* node2)
returns that the node for 0x00 is greater than the node for 0x0000, and puts it
(incorrectly at the end of the array) The comparison works like this:

1. Check if one of the nodes are 0x0 pointers
if (!node1)
    return -1;

if (!node2)
    return 1;

2. Find the shortest key length, and compare the values of the key up to this length. This
will be equal for the first byte of 0x00 and 0x0000

const UCHAR *p1 = node1->key;
const UCHAR *p2 = node2->key;
uint len = MIN(node1->keyLength, node2->keyLength);
const UCHAR *end = p1 + len;
	 
while (p1 < end)
    if (*p1++ != *p2++)
        return p1[-1] - p2[-1];

3. Check which key is the shortest, pad it to the other keys length, and compare. Because
the pad byte is 0x00, the keys will still be equal.
		
    if (len < node1->keyLength)
        {
            int n = checkTail(len, node1);
					
            if (n)
                return n;
        }
    else if (len < node2->keyLength)
    {
            int n = checkTail(len, node2);

            if (n)
                return -n;
    }

4. Finally compare the recordNumber. The recordNumber for the 0x00 node is higher, and
therefore it is considered to sort after 0x0000.
		  	 
    return node1->recordNumber - node2->recordNumber;
}

Padding with 0x00 and comparing seems to return the wrong result in
this scenario.  The array will look like this
[0x00|0x0000|0x0000|0x00], which is merged into the actual index.

The fix for this is not to pad to the length of the longer key, but
instead compare the length of the keys. This will also make sure
that the ordering of the index keys are the same for both
DeferredIndex::addNode and IndexRootPage::indexMerge. Therefore we can
also enable the ASSERT that was commented out to hide #33190. Before
commented out, this ASSERT used to hit because of different orderings.


=== added file 'mysql-test/suite/falcon/t/falcon_bug_34478.test'
Test case for #34478, test case for #33190 already exists

=== added file 'mysql-test/suite/falcon/r/falcon_bug_34478.result'
Result file for the test case

=== modified file 'storage/falcon/DeferredIndex.cpp'
Changed the comparison methods, and removed the no longer used checkTail methods

=== modified file 'storage/falcon/DeferredIndex.h'
Removed the no longer used checkTail methods

=== modified file 'storage/falcon/IndexRootPage.cpp'
Uncommented the ASSERT and removed the debug message