Bug#41360 - backup_views fails after merge of main and backup trees
The backup error stack was reordered a while ago to match the ordering of the server
error stack. When a view cannot be restored due to a missing dependency, the top-level
error ER_BACKUP_CANT_RESTORE_VIEW is now displayed, whereas the underlying problem
ER_NO_SUCH_TABLE was displayed earlier. This patch changes the backup_views test to
accommodate this reordering.