Bug#34182 - Falcon needs to be able to keep locks at lower savepoints when they are explicitly released or uinlocked at hight savepoints  Locks should be retained during the life of that savepoint unless they are explicitely unlocked before the next savepoint.
It is not necessary to add a new lock record at each savepoint.  The existing lock will do the job as long as it is not unlocked at a later savepoint.  The best solution to this problem is to only unlock records that were locked in the current savepoint, or verb.

A new integer is sent into unlockRecords which holds the current verbMark or
savepoint.  LockRecords that are at a higher savepoint are not unlocked.  Then, if a
commit, rollback, releaseSavepoint, or rollbackToSavepoint occurs that would release that
lockRecord, it will get released just like it did before.  

A new testcase is added  that shows the affect of locks attained at multiple savepoints.