BUG#32702 - Backup: DDL blocker does not block all it should
After this patch, the DDL blocker will be renamed to Backup Metadata Lock and 
will block all the statements listed in WL#4644. Test backup_ddl_blocker will be 
renamed to backup_bml and extended to test all these statements.

This patch implements desing described in WL#4644. It marks all statements which 
need to be blocked by BML with special CF_BLOCKED_BY_BML flag. Then registering 
and unregistering marked statements with BML is done globally in 
mysql_execute_command(). 

This makes changes in the parser code minimal, makes it easy to trace the list 
of statements which ale blocked and easily ensures that each bml_enter() call is 
matched with bml_leave().