Bug#35255 : test falcon_bug_22165 fails sporadically.
Problem : Crashes happen in a concurrent workload with inserts/create table/drop
table, or insert/alter table.  The reason for assertions is a race condition in Dbb::findSection() function, that is called on the first insert() after table is created ot altered. When 2 threads are looking for the same section  in Dbb::sections hashtable at the very same time and do not find it , they will  insert 2 equal entries into the table. One entry will be deleted, when table is dropped. Another entry will stay and will with high probability become invalid  when section for this table is reused (invalid means root page does not match sectionId).  In this case, next access to section will fail with an assert.
      
The fix is to synchronize access to the Dbb::section hashtable with a mutex .There are very few  calls to Dbb::findSection in normal case, so synchronization will not introduce performance losses.