Bug#42828 - The background for this code is that we had found a duplicate key value in scanIndex so we know that at least one of the record versions contain a duplicate.  We search the base record and each prior record to find any and all possible duplicate records and see if they are visible or not.  If they are pending, we should wait for that other transaction to complete.  Here, we found a record version that is active but not a duplicate.  We must need to note the activeTransaction before looking further.  If we later find a duplicate that would be visible if this transaction completed, then we must wait on that active transaction.
What happened here is that sometime between the instant that getRelativeState() was called to find that the dup->transaction was still pending and the current instant, the transaction was detached from that recordVersion.  The transaction was probably rolled back since that happens a lot quicker than a commit, writeComplete, and scavenge.

This code can just safely keep looking for dups if the transaction is returned null here.