Bug #39445 "Falcon: Update fails following online add column"
Enabled online ALTER ADD COLUMN.

The problem is that online ALTER ADD COLUMN creates a new default
record format for the table that has more fields than the older
record formats associated with existing rows.

This causes inconsistent results during row updates because the
the older, shorter rows lacked the encoding for new columns.

StorageInterface::decodeRecord() now sets the NULL flag for fields
that are not physically represented. The assumption of NULL can be
safely made because online ALTER ADD COLUMN only supports nullable
columns and a default value of NULL. 

Columns added outside of this restriction, e.g. with a non-null
default value, will be created via the slow ALTER in which all
rows are copied into a new table and thus assigned the newer
record format.