Bug #37714 rpl.rpl_heartbeat fails sporadically in pushbuild due to timeout
Logs on PB show that the IO thread was down by the clean-up (drop table t1) of the test.
A propable reason for IO thread to stop is a small value of slave_net_timeout - chosen as 
tradeoff betweeen a need to test counting of heartbeats and the test execution time.
On a slow env it can be that the timeout elapses first before any heartbeat got arrived.

Fixed with performing the clean-up separately by the master and the slave.