Bug #43048 Falcon may crash during shutdown when running it embedded
Calling shutdown() directly on a Thread object is not thread safe. If the
thread we want to shutdown wake's up during the call to shutdown() it might
shutdown and delete the thread object before the call to shutdown() has 
completed. This might result in shutdown() accessing free and overwritten
memory.

Instead of calling shutdown() directly, use the thread barn's
shutdown() method which has the neccessary locking for concurrent
threads.